third_party/WebKit/LayoutTests/printing/multicol-not-supported-expected.txt

PASS internals.numberOfPages(800, 800) is 6
PASS successfullyParsed is true

TEST COMPLETE
To run this test manually, print this document.

There should be 6 pages in this document. Multicol doesn't work well with printing, so it should be disabled. See crbug.com/99358
